Launch properly with SSL; require -port > 0
The previous version tried to allow -port=0 as a valid option so that a
listening port could be allocated dynamically and logged. However, using
http.Server.ListenAndServeTLS() doesn't provide a way to extract the dynamic
address.

Rather than reimplement parts of http.Server.ListenAndServeTLS(), the -port
argument is no longer optional. This has the benefit of allowing the use of
http.server.ListenAndServe() in non-SSL mode as well, eliminating the need for
manual net.Listener setup.